refactor(api): move target profile for creation to prescription/target-profile subdomain
