+++ api/lib/domain/usecases/create-target-profile.js
@@ -1,5 +1,5 @@
+import { TargetProfileForCreation } from '../../../src/prescription/target-profile/domain/models/TargetProfileForCreation.js';
 import { TargetProfileCannotBeCreated } from '../../../src/shared/domain/errors.js';
-import { TargetProfileForCreation } from '../../../src/shared/domain/models/TargetProfileForCreation.js';
 
 const createTargetProfile = async function ({
   targetProfileCreationCommand,
