+++ api/tests/tooling/domain-builder/factory/build-target-profile-for-creation.js
@@ -1,4 +1,4 @@
-import { TargetProfileForCreation } from '../../../../src/shared/domain/models/TargetProfileForCreation.js';
+import { TargetProfileForCreation } from '../../../../src/prescription/target-profile/domain/models/TargetProfileForCreation.js';
 
 const buildTargetProfileForCreation = function ({
   name = 'Profil cible super cool',
